requirements.txt
openai>=1.0.0          # For OpenAI API client
pandas>=2.0.0          # Used in md_to_summary.py for data handling
python-dotenv>=0.19.0  # Recommended for handling API keys 
python-docx